gitanalysis/requirement.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
python-gitlab==2.0.1
requests==2.22.0
urllib3==1.25.8
